Rename to grapl_plugin_sdk.analyzer
