+++ pants.toml
@@ -68,7 +68,7 @@ root_patterns = [
   "/pulumi",
   "/src/proto",
   "/src/python/graphql_endpoint_tests",
-  "/src/python/grapl-analyzerlib-new",
+  "/src/python/grapl-plugin-sdk",
   "/src/python/grapl-common",
   "/src/python/graplctl",
   "/src/python/grapl-tests-common",
