--- src/python/grapl-analyzerlib-new/BUILD
@@ -1,11 +1,0 @@
-python_distribution(
-    dependencies=[
-        "./grapl_analyzerlib_new",
-        "./grapl_analyzerlib_new:py_typed",
-    ],
-    provides=setup_py(
-        name="grapl-analyzerlib-new",
-        version="0.0.1",
-        description="Library for Grapl Analyzers",
-    ),
-)
