+++ src/python/grapl-plugin-sdk/BUILD
@@ -1,0 +1,11 @@
+python_distribution(
+    provides=setup_py(
+        name="grapl-plugin-sdk",
+        version="0.0.1",
+        description="Library for developing Grapl Plugins, like Analyzers.",
+    ),
+    dependencies=[
+        "./grapl_plugin_sdk",
+        "./grapl_plugin_sdk:py_typed",
+    ],
+)
