+++ src/python/grapl-plugin-sdk/grapl_plugin_sdk/analyzer/BUILD
@@ -1,0 +1,1 @@
+python_sources()
