+++ src/python/grapl-plugin-sdk/grapl_plugin_sdk/py.typed
